# app/Http/Controllers/API/v1/UsersController.php
<?php

namespace App\Http\Controllers\API\v1;

use App\Models\User;
use App\Http\Controllers\Controller;
use App\Http\Handlers\UserHandler;
use App\Http\Resources\UserResource;
use Illuminate\Http\Request;

class UsersController extends Controller
{
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function index()
    {
        //
    }

    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     */
    public function store(Request $request)
    {
        //
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function show($id)
    {
        //
    }

    /**
     *
     * @OA\PATCH(
     *     path="/api/v1/users/{id}",
     *     @OA\Parameter(
     *         name="gender",
     *         description="Gender",
     *         in="query",
     *         @OA\Schema(
     *             type="string"
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="age_group",
     *         description="age group",
     *         in="query",
     *         @OA\Schema(
     *             type="string"
     *         )
     *     ),    
     *     
     *   @OA\Response(response="200", description="UserResource")
     * )
     *
     * Update the specified resource in storage.
     *
     * @param  int $id
     * @param  \Illuminate\Http\Request $request
     * @param UserHandler $userHandler
     * @return \Illuminate\Http\Response
     * @throws \Illuminate\Validation\ValidationException
     */
    public function update($id, Request $request, UserHandler $userHandler)
    {
        $this->validate($request, [
            'gender'    => 'sometimes|string',
            'age_group' => 'sometimes|string',
        ]);

        $userId = $request->id;
        $transformedUser = new UserResource($userHandler->create($userId, $request));

        return response($transformedUser, 200);
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function destroy($id)
    {
        //
    }
}