# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "summary": "Returns all categories.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\CategoriesController::index",
                "parameters": [
                    {
                        "name": "CategoryService",
                        "in": "query",
                        "description": "CategoryService Object",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all categories"
                    }
                }
            },
            "put": {
                "summary": "Update a category.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\CategoriesController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Uuid of the category",
                        "required": true,
                        "schema": {
                            "type": "char(36)"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "New name for the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "Icon of the category to replace existing one",
                        "required": false,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated"
                    },
                    "400": {
                        "description": "Category not found"
                    }
                }
            },
            "post": {
                "summary": "Create and save a category.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\CategoriesController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "Icon of the category",
                        "required": false,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created"
                    }
                }
            },
            "delete": {
                "summary": "Delete a category and its related icon.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\CategoriesController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Uuid of the category to delete",
                        "required": true,
                        "schema": {
                            "type": "char(36)"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated"
                    },
                    "400": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/api/v1/user-businesses": {
            "post": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\UserBusinessesController::store",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business"
                    }
                }
            },
            "delete": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\UserBusinessesController::delete",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business"
                    }
                }
            }
        },
...
        "/api/v1/users/{id}": {
            "patch": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\UsersController::update",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "age_group",
                        "in": "query",
                        "description": "age group",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserResource"
                    }
                }
            }
        }
    }
}